// src/types/materialConsumptionList.gen.ts
// Code generated by cmd/codegen. DO NOT EDIT.

export interface MaterialConsumptionListDTO {
	id: number;
	date: string;
	construction_site_id: number;
	comment: string | null;
	version: number;
	construction_site: Record<string, unknown>;
}

export interface MaterialConsumptionList {
	id: number;
	date: Date;
	construction_site_id: number;
	comment: string | null;
	version: number;
	construction_site: Record<string, unknown>;
}