// src/types/materialTransferList.gen.ts
// Code generated by cmd/codegen. DO NOT EDIT.

export interface MaterialTransferListDTO {
	id: number;
	date: string;
	source_construction_site_id: number;
	destination_construction_site_id: number;
	comment: string | null;
	version: number;
	source_construction_site: Record<string, unknown>;
	destination_construction_site: Record<string, unknown>;
}

export interface MaterialTransferList {
	id: number;
	date: Date;
	source_construction_site_id: number;
	destination_construction_site_id: number;
	comment: string | null;
	version: number;
	source_construction_site: Record<string, unknown>;
	destination_construction_site: Record<string, unknown>;
}